requirements.txt
boto3==1.26.118
botocore==1.29.118
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
cryptography==40.0.2
DateTime==5.1
et-xmlfile==1.1.0
greenlet==2.0.2
idna==3.4
jmespath==1.0.1
Naked==0.1.32
ndg-httpsclient==0.5.1
numpy==1.24.2
openpyxl==3.1.2
pandas==2.0.0
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
pyDes==2.0.1
pyodbc==4.0.39
pyOpenSSL==23.1.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
shellescape==3.8.1
six==1.16.0
SQLAlchemy==2.0.10
timedelta==2020.12.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
zope.interface==6.0
